Pull request remarks
Former-commit-id: daa4b495ce0ec142a8b9ba2405ab57656c5ab39d

+++ Sources/Internal/Platform/SystemTimer.cpp
@@ -180,7 +180,7 @@ float32 SystemTimer::ElapsedSec()
 	}
 	else
 	{
-		float32 currentTime = (float32)(GetTickCount64() / 1000.0f);
+		float32 currentTime = GetTickCount64() / 1000.0f;
 		Logger::FrameworkDebug("delta %f", currentTime - t0);
 		return currentTime - t0;
 	}
